Restructure acquisition into clientkit and core modules. - update producer and consumers.
