+++ acquisition/clientKit/src/main/java/datatypes/ProducerData.java
@@ -1,0 +1,7 @@
+package datatypes;
+
+public class ProducerData {
+
+
+
+}
