+++ acquisition/clientKit/src/main/java/datatypes/RawType.java
@@ -1,0 +1,22 @@
+package datatypes;
+
+public class RawType {
+    int id;
+    int name;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public int getName() {
+        return name;
+    }
+
+    public void setName(int name) {
+        this.name = name;
+    }
+}
