+++ acquisition/clientKit/src/main/java/datatypes/RequestStatus.java
@@ -1,0 +1,9 @@
+package datatypes;
+
+public enum RequestStatus {
+    SUCCESS,
+    ERROR,
+    INPROGRESS,
+    NEW,
+    SCHEDULED
+}
