+++ acquisition/core/src/main/java/acquisition/utils/ScheduledStopTask.java
@@ -1,0 +1,18 @@
+package acquisition.utils;
+
+import java.util.TimerTask;
+
+public class ScheduledStopTask extends TimerTask {
+
+
+    TimerTask task;
+
+    public ScheduledStopTask(TimerTask task){
+        this.task = task;
+    }
+
+    @Override
+    public void run() {
+        task.cancel();
+    }
+}
