+++ acquisition/core/src/main/resources/acquisition.properties
@@ -1,0 +1,7 @@
+bootstrap.servers=localhost:9092
+acks=all
+block.on.buffer.full=true
+key.serializer=org.apache.kafka.common.serialization.StringSerializer
+value.serializer=acquisition.kafka.ObjectSerializer
+key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
+value.deserializer=acquisition.kafka.ObjectDeSerializer
