--- acquisition/src/main/java/com/uci/perpetualdb/acquisition/datatypes/ProducerData.java
@@ -1,7 +1,0 @@
-package com.uci.perpetualdb.acquisition.datatypes;
-
-public class ProducerData {
-
-
-
-}
