--- acquisition/src/main/java/com/uci/perpetualdb/acquisition/datatypes/RawType.java
@@ -1,22 +1,0 @@
-package com.uci.perpetualdb.acquisition.datatypes;
-
-public class RawType {
-    int id;
-    int name;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public int getName() {
-        return name;
-    }
-
-    public void setName(int name) {
-        this.name = name;
-    }
-}
