--- acquisition/src/main/java/com/uci/perpetualdb/acquisition/datatypes/RequestStatus.java
@@ -1,9 +1,0 @@
-package com.uci.perpetualdb.acquisition.datatypes;
-
-public enum RequestStatus {
-    SUCCESS,
-    ERROR,
-    INPROGRESS,
-    NEW,
-    SCHEDULED
-}
