--- acquisition/src/main/java/com/uci/perpetualdb/acquisition/requestmanagement/RequestManager.java
@@ -1,23 +1,0 @@
-package com.uci.perpetualdb.acquisition.requestmanagement;
-
-import com.uci.perpetualdb.acquisition.datatypes.Request;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class RequestManager {
-
-
-    private List<Request> requests = new ArrayList<>();
-
-    protected boolean addRequest(Request request){
-        requests.add(request);
-
-        return true;
-    }
-
-    protected boolean addRequests(List<Request> newRequests){
-        requests.addAll(newRequests);
-        return true;
-    }
-}
