--- acquisition/src/main/resources/acquisition.properties
@@ -1,7 +1,0 @@
-bootstrap.servers=localhost:9092
-acks=all
-block.on.buffer.full=true
-key.serializer=org.apache.kafka.common.serialization.StringSerializer
-value.serializer=com.uci.perpetualdb.acquisition.kafka.ObjectSerializer
-key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
-value.deserializer=com.uci.perpetualdb.acquisition.kafka.ObjectDeSerializer
