--- acquisition/src/main/resources/log4j2.xml
@@ -1,18 +1,0 @@
-<Configuration status="ERROR">
-    <Appenders>
-        <Console name="stdout" target="SYSTEM_OUT">
-            <PatternLayout pattern="[%d] (%c) - %p %m %n"/>
-        </Console>
-    </Appenders>
-    <Loggers>
-        <Root level="error">
-            <AppenderRef ref="stdout"/>
-        </Root>
-        <Logger name="org.apache.kafka.clients" level="warn" additivity="false">
-            <AppenderRef ref="stdout"/>
-        </Logger>
-        <Logger name="com.uci.perpetualdb.acquisition" level="trace" additivity="false">
-            <AppenderRef ref="stdout"/>
-        </Logger>
-    </Loggers>
-</Configuration>
